Move the form classes for image and date in the frontend into their own file
